Add files via upload
This C++ program sorts an array in both ascending and descending order. It first takes the number of elements and the array values as input. The program uses nested loops to sort the array in ascending order, displays the result, and then repeats the process to sort it in descending order. Finally, both sorted arrays are printed.